+++ TagsCloud/TagsCloud/Infrastructure/PictureBoxImageHolder.cs
@@ -3,8 +3,10 @@ using System.Drawing;
 using System.Drawing.Imaging;
 using System.Linq;
 using System.Windows.Forms;
+using TagsCloud.Layouters;
+using TagsCloud.WordsProcessing;
 
-namespace TagsCloud
+namespace TagsCloud.Infrastructure
 {
     public class PictureBoxImageHolder : PictureBox, IImageHolder
     {
